third commit

+++ GasStationOnly/Models/GassStation.cs
@@ -10,15 +10,16 @@ namespace GasStationOnly.Models
     {
         public int GassStationId { get; set; }
         public string Gasoline { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public DateTime Time { get; set; }
-        public int Amount { get; set; }
-        public int Sum { get; set; }
-        public int PrePrice { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Sum { get; set; }
+        public decimal PrePrice { get; set; }
        
         public override string ToString()
         {
             return $"{Amount} - {PrePrice} - {Time}";
         }
+        
     }
 }
